requirements.txt
numpy
flask
langchain-community
langchain_groq
sentence_transformers
faiss-cpu
keyboard
